Add info helpers
When dealing with integration, one must adapt this strategy. This commit
aims to help this integration by providing some default values in info.

Note that whether the developer asked for or the user gives these values
is of no importance here, the values will just be nil - which is the
value of a non-existing key in an array.